prepare for embed
